Updated thesaurus
Thesaurus is now loaded in by a thesaurus.js file. This is currently a work around since file support is poor in web extensions. There is some documentation on the issue but the solution is not clear currently. The python script create-javascript.py currently generates the needed "thesaurus.js" javascript file. After created the file gets moved to "../plugin/thesaurus.js".