Hello world we have dom manipulation
We can now send the text to the background and apply transformations. Text is then changed and sent to the foreground to allow it to apply transformations to the dom. Current pitfalls are that elements have to have ids, this can likely be mitigated later with using indexdb or something similar.